fix(renderer): rename MessagetText.tsx to MessageText.tsx
Fix typo in filename — extra 't' in MessagetText. Updated all import
references in MessageList.tsx and related test files.
