+++ tests/unit/formatMessageTime.test.ts
@@ -1,5 +1,5 @@
 import { describe, it, expect, vi, afterEach } from 'vitest';
-import { formatMessageTime } from '@renderer/pages/conversation/Messages/components/MessagetText';
+import { formatMessageTime } from '@renderer/pages/conversation/Messages/components/MessageText';
 
 describe('formatMessageTime', () => {
   afterEach(() => {
